to-do labels and bug fixing
